Static Range Sum Queries (int to long)

+++ Hello.java
@@ -5,21 +5,21 @@ class Hello{
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
         int m = sc.nextInt();
-        int[] arr = new int[n];
+        long[] arr = new long[n];
         for(int i=0;i<n;i++){
-            arr[i] = sc.nextInt();
+            arr[i] = sc.nextLong();
         }
         inputHandler(m, sc, arr);
     }
-    public static void inputHandler(int m,Scanner sc,int[] arr){
+    public static void inputHandler(int m,Scanner sc,long[] arr){
         for(int i=0;i<m;i++){
             int st = sc.nextInt();
             int ed = sc.nextInt();
             System.out.println(prefixSum(arr, st, ed));
         }
     }
-    public static int prefixSum(int[] arr,int st,int ed){
-        int sum =0;
+    public static long prefixSum(long[] arr,int st,int ed){
+        long sum =0;
         for(int i=st-1;i<ed;i++){
             sum += arr[i];
         }
